Starting progress on validating the R2RML mapping with OnTop.

+++ src/main/java/swiss/sib/swissprot/r2s2/optimization/TableMergingConcurence.java
@@ -13,6 +13,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -41,25 +42,21 @@ public class TableMergingConcurence {
 
 	public List<Table> run() {
 		logger.info("Starting merging tables");
-		List<Table> mergeCandidates = tables.stream().filter(this::hasDistinctSubjects)
+		List<Table> mergeCandidates = tables.stream().filter(Predicate.not(this::hasDistinctSubjects))
 				.sorted((a, b) -> compareTablesFirstTypeThenSize(a, b)).collect(Collectors.toList());
-		Iterator<Table> mergeCandidatesIter = mergeCandidates.iterator();
-		while (mergeCandidatesIter.hasNext()) {
-			Table mergeCandidate = mergeCandidatesIter.next();
-			if (!isTableEmpty(mergeCandidate)) {
-				Iterator<Table> overlappingSubjects = findTablesWithOverlappingSubjects(mergeCandidate,
-						List.copyOf(mergeCandidates));
-				while (overlappingSubjects.hasNext()) {
-					final Table other = overlappingSubjects.next();
-					merge(mergeCandidate, other);
-					if (isTableEmpty(other)) {
-						overlappingSubjects.remove();
+		Iterator<Table> mcIter = mergeCandidates.iterator();
+		while (mcIter.hasNext()) {
+			Table mc = mcIter.next();
+			if (!isTableEmpty(mc)) {
+				for (Table other : List.copyOf(mergeCandidates)) {
+					if (haveOverlappingSubjects(mc, other)) {
+						merge(mc, other);
 					}
 				}
+				mergedTables.add(mc);
 			} else {
-				mergeCandidatesIter.remove();
+				mcIter.remove();
 			}
-			mergedTables.add(mergeCandidate);
 		}
 		removeEmpty(tables);
 		logger.info("Finished merging tables");
@@ -76,10 +73,14 @@ public class TableMergingConcurence {
 	private int compareTablesFirstTypeThenSize(Table a, Table b) {
 		if (a.name().startsWith("type_") && !b.name().startsWith("type_"))
 			return -1;
-		else if (!a.name().startsWith("type_") && !a.name().startsWith("type_"))
+		else if (!a.name().startsWith("type_") && b.name().startsWith("type_"))
 			return 1;
 		else {
-			return Long.compare(size(a, false), size(b, false));
+			final int sizeComparison = Long.compare(size(a, false), size(b, false));
+			if (sizeComparison == 1) {
+				return a.name().compareTo(b.name());
+			}
+			return sizeComparison;
 		}
 	}
 
@@ -87,7 +88,7 @@ public class TableMergingConcurence {
 		Iterator<Table> ti = tables.iterator();
 		while (ti.hasNext()) {
 			Table t = ti.next();
-			if (isTableEmpty(t)) {
+			if (isTableEmpty(t) && !allColumnsVirtual(t)) {
 				try (Connection conn = open(path); var stat = conn.createStatement()) {
 					String dropSql = "DROP TABLE " + t.name();
 					logger.info("Running " + dropSql);
@@ -98,6 +99,21 @@ public class TableMergingConcurence {
 				}
 			}
 		}
+	}
+
+	/**
+	 * Avoid dropping tables that have all virtual columns.
+	 */
+	private boolean allColumnsVirtual(Table t) {
+		for (Column c : t.subject().getColumns())
+			if (c.isPhysical())
+				return false;
+		for (PredicateMap p : t.objects()) {
+			for (Column c : p.columns().getColumns())
+				if (c.isPhysical())
+					return false;
+		}
+		return true;
 	}
 
 	public boolean isTableEmpty(Table t) {
@@ -148,12 +164,12 @@ public class TableMergingConcurence {
 
 	public void merge(Table mc, Table other, PredicateMap pm, List<Column> toMerge, Connection conn, Statement stat)
 			throws SQLException {
-		String msc = concatColumns(mc, "mc");
-		String osc = concatColumns(other, "oc");
+		String msc = concatSubjectColumns(mc, "mc");
+		String osc = concatSubjectColumns(other, "oc");
 
 		String update = "UPDATE " + mc.name() + " SET "
 				+ toMerge.stream().map(oc -> "" + oc.name() + " = oc." + oc.name()).collect(Collectors.joining(" , "))
-				+ " FROM " + other.name() + " oc WHERE " + concatColumns(mc, "") + '=' + osc;
+				+ " FROM " + other.name() + " oc WHERE " + concatSubjectColumns(mc, "") + '=' + osc;
 
 		String delete = "DELETE FROM " + other.name() + " oc USING " + mc.name() + " mc WHERE " + msc + '=' + osc;
 
@@ -167,42 +183,38 @@ public class TableMergingConcurence {
 		size(other, true);
 	}
 
-	private Iterator<Table> findTablesWithOverlappingSubjects(Table mc, List<Table> tables) {
-		List<Table> compatible = new ArrayList<>();
-		for (Table other : tables) {
-			final boolean sameSubjectKind = mc.subjectKind().equals(other.subjectKind());
-			Set<IRI> mcPredicates = mc.objects().stream().map(PredicateMap::predicate).collect(Collectors.toSet());
-			Set<IRI> otherPredicates = other.objects().stream().map(PredicateMap::predicate)
-					.collect(Collectors.toSet());
-			final Stream<IRI> filter = mcPredicates.stream().filter(otherPredicates::contains);
-			final boolean samePredicate = filter.findAny().isPresent();
-			if (other != mc && sameSubjectKind && !samePredicate && !mergedTables.contains(other)
-					&& !isTableEmpty(other)) {
-				try (Connection conn = open(path); Statement statement = conn.createStatement()) {
-					String msc = concatColumns(mc, "mc");
-					String osc = concatColumns(other, "oc");
-					String sql = "SELECT COUNT(*) FROM " + mc.name() + " mc , " + other.name() + " oc WHERE " + msc
-							+ '=' + osc;
-					logger.info("Running " + sql);
-					try (ResultSet rs = statement.executeQuery(sql)) {
-						rs.next();
-						final boolean isAMergeCandidate = rs.getInt(1) > 0;
-						if (isAMergeCandidate) {
-							logger.info(mc.name() + " is a merge candidate with " + other.name() + " overlapping ");
-							compatible.add(other);
-						} else {
-							logger.info(mc.name() + " is NOT a merge candidate with " + other.name());
-						}
+	public boolean haveOverlappingSubjects(Table mc, Table other) {
+		final boolean sameSubjectKind = mc.subjectKind().equals(other.subjectKind());
+		Set<IRI> mcPredicates = mc.objects().stream().map(PredicateMap::predicate).collect(Collectors.toSet());
+		Set<IRI> otherPredicates = other.objects().stream().map(PredicateMap::predicate).collect(Collectors.toSet());
+		final Stream<IRI> filter = mcPredicates.stream().filter(otherPredicates::contains);
+		final boolean samePredicate = filter.findAny().isPresent();
+		if (other != mc && sameSubjectKind && !samePredicate && !mergedTables.contains(other) && !isTableEmpty(other)) {
+			try (Connection conn = open(path); Statement statement = conn.createStatement()) {
+				String msc = concatSubjectColumns(mc, "mc");
+				String osc = concatSubjectColumns(other, "oc");
+				String sql = "SELECT COUNT(*) FROM " + mc.name() + " mc , " + other.name() + " oc WHERE " + msc + '='
+						+ osc;
+				logger.info("Running " + sql);
+				try (ResultSet rs = statement.executeQuery(sql)) {
+					rs.next();
+					final boolean isAMergeCandidate = rs.getInt(1) > 0;
+					if (isAMergeCandidate) {
+						logger.info(mc.name() + " is a merge candidate with " + other.name() + " overlapping ");
+						return true;
+					} else {
+						logger.info(mc.name() + " is NOT a merge candidate with " + other.name());
+						return false;
 					}
-				} catch (SQLException e) {
-					throw new IllegalStateException(e);
 				}
+			} catch (SQLException e) {
+				throw new IllegalStateException(e);
 			}
 		}
-		return compatible.iterator();
+		return false;
 	}
 
-	public String concatColumns(Table mc, String alias) {
+	public String concatSubjectColumns(Table mc, String alias) {
 		return mc.subject().getColumns().stream().map(c -> {
 			if (c.isPhysical()) {
 				if (alias.isEmpty()) {
@@ -223,17 +235,16 @@ public class TableMergingConcurence {
 			return false;
 		}
 		try (Connection conn = open(path); Statement statement = conn.createStatement()) {
-			String sql = "SELECT " + sc + " FROM " + t.name() + " GROUP BY " + sc + " HAVING (COUNT(DISTINCT(" + sc
-					+ ")) > 1) LIMIT 1";
+			String sql = "SELECT " + sc + " FROM " + t.name() + " GROUP BY " + sc + " HAVING (COUNT(*) > 1) LIMIT 1";
 			logger.info("Running " + sql);
 			try (ResultSet rs = statement.executeQuery(sql)) {
-				final boolean isNotAMergeCandidate = rs.next();
-				if (isNotAMergeCandidate) {
-					logger.info(t.name() + " is NOT a merge candidate");
-				} else {
+				final boolean isAMergeCandidate = rs.next();
+				if (isAMergeCandidate) {
 					logger.info(t.name() + " is a merge candidate");
+				} else {
+					logger.info(t.name() + " is NOT a merge candidate");
 				}
-				return !isNotAMergeCandidate;
+				return isAMergeCandidate;
 			}
 		} catch (SQLException e) {
 			throw new IllegalStateException(e);
